requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
install==1.3.5
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
whitenoise==6.4.0
